backend/requirements.txt
Django>=6.0.1
djangorestframework
django-cors-headers
djangorestframework-simplejwt
django-environ
Pillow
